add Telegrams helper class
to parse and convert Telegram only once instead of once per device trigger
